+++ src/app/wines/[id]/page.scss
@@ -1,0 +1,3 @@
+.page {
+    padding: 20px 16px;
+}
