+++ src/components/cardComponent/Soohyun.scss
@@ -6,3 +6,5 @@
         width: 58px;
     }
 }
+
+//동규님 scss 작업 다 끝내셨으면 이 파일 지우시면 됩니당
