+++ src/components/ratingbar/ratingBar.tsx
@@ -16,7 +16,7 @@ interface RatingBarProps {
 }
 
 const RatingBar: React.FC<RatingBarProps> = ({ score, avgRatings, reviewCount }) => {
-    const percentage = (avgRatings[score] / reviewCount) * 100;
+    const percentage = reviewCount && (avgRatings[score] / reviewCount) * 100;
     console.log(percentage);
 
     return (
